Update docs to require FCM server key for standalone Android apps
The SNS/GCM system is less reliable than FCM and less maintained and we have announced that we are requiring FCM for all new standalone apps: https://blog.expo.io/upcoming-changes-to-the-expo-push-notification-service-summer-2018-254030da786b. This commit updates the push notification guide and adjusts the wording on the FCM guide to reflect this change.

fbshipit-source-id: 7e76bed