add continuous integration guide with example code
Sorry, it took me a while to put this together, but here it is!

I created a slightly more unopinionated version of my earlier [article on Medium](https://blog.expo.io/setting-up-expo-and-bitbucket-pipelines-8995ef036a18). The documentation also covers the configuration, example repositories and actual pipeline logs/views for Travis, Gitlab and Bitbucket Pipelines. I hope this helps!

For the documentation blocks, I used the `details` and `summary` HTML elements. Without this, I think inexperienced developers are a bit overwhelmed by the different chunks of configuration. I'm not sure if I should create some custom components to improve the styling and make it more Expo-themed. I'd be happy to still do that if you guys want that!

Hope it helps!

Cheers,
Cedric ❤️

Closes expo/expo-docs#340

fbshipit-source-id: 6fc19dd