Rename to expo-server-sdk-*; delete internal Ruby SDK (OSS-first now)
This renames the server SDKs from exponent-server-sdk-* to expo-server-sdk-*, including the shipit configs for the Node and Rust SDKs. The Ruby SDK is community-maintained, so this commit removes Expo's internal copy of it and removes it from the shipit config as well. The docs are all updated with the latest links to the newly renamed GitHub repos.

Before landing this commit, I'll re-deploy the shipit service and apply the Terraform changes.

fbshipit-source-id: a7153cf