Added CircleCI sample configuration to 'setting-up-continuous-integration.md'
First of all, thanks for your great work on Expo (docs). It really saved us a ton of work (and time)!

We were really struggling to get the automated publishing (Expo) to work on CircleCI. The process did sometimes succeed but most of the time it was -just- hanging at 100%. The trick was to set `--max-workers 1` and now it works fine.

I've added CircleCI sample configuration for automated `expo publish` to 'setting-up-continuous-integration.md' to help other devs set this up in the future. I hope you like it!

Please let me know if you have some feedback.

Closes expo/expo-docs#362

fbshipit-source-id: 5b6226b