Note regarding default Permissions
Note that the list of permissions listed at the end of document are default permissions which will end up in the standalone Android build if `android.permissions` is not set in `app.json`

<!--
Thanks for helping! Please check that you have edited the docs in the
`versions/unversioned` directory, if you want these changes to apply to
the next SDK version too.
-->

Closes expo/expo-docs#364

fbshipit-source-id: d13425a