Fix typo and sendSMSAsync return result
In the return result for sendSMSAsync, it states that canceling
an SMS returns "{ result: canceled }" when in reality after
testing, it returns "{ result: cancelled }". This fix makes the
docs reflect the current behavior of the code, regardless of
which spelling was intended.

<!--
Thanks for helping! Please check that you have edited the docs in the
`versions/unversioned` directory, if you want these changes to apply to
the next SDK version too.
-->

Closes expo/expo-docs#389

fbshipit-source-id: ac93c34